+++ packages/editor/editor-core/src/__tests__/unit/plugins/block-type/keymap.ts
@@ -15,16 +15,30 @@ import {
   tr,
   tdEmpty,
   tdCursor,
+  simulatePlatform,
+  Platforms,
 } from '@atlaskit/editor-test-helpers';
 import { analyticsService } from '../../../../analytics';
 import { setNodeSelection } from '../../../../utils';
+import { CreateUIAnalyticsEventSignature } from '@atlaskit/analytics-next-types';
+import {
+  AnalyticsEventPayload,
+  INPUT_METHOD,
+  ACTION_SUBJECT_ID,
+  EVENT_TYPE,
+  ACTION_SUBJECT,
+  ACTION,
+} from '../../../../plugins/analytics';
+import { EditorView } from 'prosemirror-view';
 
 describe('codeBlock - keymaps', () => {
   const createEditor = createEditorFactory();
+  let createAnalyticsEvent: CreateUIAnalyticsEventSignature;
 
   let trackEvent;
-  const editor = (doc: any) =>
-    createEditor({
+  const editor = (doc: any) => {
+    createAnalyticsEvent = jest.fn(() => ({ fire() {} }));
+    return createEditor({
       doc,
       editorProps: {
         analyticsHandler: trackEvent,
@@ -33,14 +47,16 @@ describe('codeBlock - keymaps', () => {
         allowLists: true,
         allowTables: true,
         allowRule: true,
+        allowAnalyticsGASV3: true,
       },
+      createAnalyticsEvent,
     });
+  };
 
   beforeEach(() => {
     trackEvent = jest.fn();
     analyticsService.trackEvent = trackEvent;
   });
-
   describe('keymap', () => {
     describe('when hits cmd-z', () => {
       it('should undo last autoformatting', () => {
@@ -398,6 +414,44 @@ describe('codeBlock - keymaps', () => {
               'atlassian.editor.movedown.keyboard',
             );
           });
+        });
+      });
+    });
+
+    describe('when hits Cmd-Alt-9', () => {
+      describe('mac', () => {
+        simulatePlatform(Platforms.Mac);
+        let editorView: EditorView;
+        beforeEach(() => {
+          ({ editorView } = editor(doc(p('{<}text{>}'))));
+
+          sendKeyToPm(editorView, 'Cmd-Alt-9');
+        });
+
+        it('should toggle block quotes', () => {
+          expect(editorView.state.doc).toEqualDocument(
+            doc(blockquote(p('text'))),
+          );
+        });
+
+        it('should create Analytics GAS V3 event', () => {
+          const expectedPayload: AnalyticsEventPayload = {
+            action: ACTION.FORMATTED,
+            actionSubject: ACTION_SUBJECT.TEXT,
+            eventType: EVENT_TYPE.TRACK,
+            actionSubjectId: ACTION_SUBJECT_ID.FORMAT_BLOCK_QUOTE,
+            attributes: {
+              inputMethod: INPUT_METHOD.KEYBOARD,
+            },
+          };
+
+          expect(createAnalyticsEvent).toHaveBeenCalledWith(expectedPayload);
+        });
+
+        it('should track event for Analytics v2', () => {
+          expect(trackEvent).toHaveBeenCalledWith(
+            'atlassian.editor.format.blockquote.keyboard',
+          );
         });
       });
     });