+++ packages/navigation/atlassian-switcher/src/components/atlassian-switcher.tsx
@@ -2,17 +2,19 @@ import * as React from 'react';
 import JiraSwitcher from './jira-switcher';
 import ConfluenceSwitcher from './confluence-switcher';
 
+interface AtlassianSwitcherProps {
+  product: string;
+  cloudId: string;
+  triggerXFlow: Function;
+}
+
 export default ({
   product,
   cloudId,
   triggerXFlow,
   ...props
-}: {
-  product: string;
-  cloudId: string;
-  triggerXFlow: Function;
-}) => {
-  let Switcher: React.ReactType = JiraSwitcher;
+}: AtlassianSwitcherProps) => {
+  let Switcher: React.ReactType;
   switch (product) {
     case 'jira':
       Switcher = JiraSwitcher;
@@ -21,7 +23,13 @@ export default ({
       Switcher = ConfluenceSwitcher;
       break;
     default:
-      Switcher = JiraSwitcher;
+      if (process.env.NODE_ENV !== 'production') {
+        // tslint:disable-next-line:no-console
+        console.warn(
+          `Product key ${product} provided to Atlassian Switcher doesn't have a corresponding product specific implementation.`,
+        );
+      }
+      return null;
   }
   return <Switcher cloudId={cloudId} triggerXFlow={triggerXFlow} {...props} />;
 };